Flesh out ROS 2 support (except for lightweight_ur_interface) (#19)
* Style & consistency fixes

* Support for split ROS 1/ROS 2 in tri_hardware_drivers metapackage

* Support for split ROS 1/ROS 2 in tri_mocap_common

* Make build options consistent with CRU, add scripts to select between ROS 1 and ROS 2

* Restructure ROS 2 node init calls to not remove ROS args

* Use C++17 to match latest rclcpp

+++ schunk_wsg_driver/src/schunk_wsg_driver/schunk_wsg_driver_common.cpp
@@ -406,7 +406,8 @@ WSGRawCommandMessage WSGInterface::MakePrePositionCommand(
     double speed_mm_per_s) const
 {
   WSGRawCommandMessage preposition_command(kPrePosition);
-  const uint8_t flags = stop_mode | move_mode;
+  const uint8_t flags =
+      static_cast<uint8_t>(stop_mode) | static_cast<uint8_t>(move_mode);
   preposition_command.AppendParameterToBuffer(flags);
   preposition_command.AppendParameterToBuffer(static_cast<float>(width_mm));
   preposition_command.AppendParameterToBuffer(
